requirements.txt
fastapi
uvicorn
gradio
diffusers
torch
transformers
Pillow
grpcio
grpcio-tools
